fix: metro config not overriden on newer Expo 54 (#1676)
Fixes `metro-config` import not being overridden correctly on new Expo
54 versions.
The issue was caused by `metro-config` being built in a different way
which changed the shape of exported object from
```js
{
  loadConfig: [async function],
  /* etc... */
}
```
to
```js
{
  default: {
    loadConfig: [async function],
    /* etc... */
  },
  loadConfig: [Getter],
  /* etc... */
}
```
which broke our override.

This PR changes the implementation to instead override the whole export
with a `Proxy` object which correctly forwards `loadConfig` calls to our
implementation.

### How Has This Been Tested: 
- update dependencies in `expo54` test app to 
```
    "expo": "^54.0.0",
```
- install with `npm install && npx expo install --fix`
- launch app
- it should launch