Enable Radon AI by default after redoing MCP server handling logic to use newly available APIs (#1665)
This PR switches Radon AI on by default. We previously had to disable
Radon AI on Cursor, as we couldn't find a good way to register the
server without triggering some annoying modals or modifying files that
belong to the users' workspace.

Recently Cursor added MCP integration API for extensions and this PR
adapts it along with implementing a couple of improvements over the
current mechanism:
https://cursor.com/docs/context/mcp-extension-api#api-reference

While working on this update we discovered several issues with the way
session/error handling was implemented in the local MCP server. This PR
updates these along the way as they turned out to be quite coupled with
the rest of the changes.



The key highlights of changes made here:
1) We update the code resposible for managing the mcp.json files. We no
longer add new entries to these files and the lefover code is here to
cleanup entries that the previous versions of Radon may have written. We
make sure only http entries for Radon are deleted as now people can also
add stdio entry to integrate Radon with other MCP clients or older
versions of Cursor/VSCode.
2) The local http server now supports session properly. We use the
template from
https://github.com/modelcontextprotocol/typescript-sdk?tab=readme-ov-file#with-session-management
for proper session id generation and for handling lost sessions. This
helps with the scenarios where you have multiple agents connected (i.e.
VSCode and Claude Code)
3) We delete all the reload handling logic that'd result in server being
restarted. All MCP clients like to keep mcp sessions active and when
restarting they need to go over initialization again. In most cases when
MCP server gets disconnected, the client wouldn't retry the connection
again. This is now changed and we avoid restarting the server altogether
unless Radon AI is disabled in the settings.
4) Instead of restarting we now rely on tools-changed notification. This
is a required part of MCP protocol. Apparently some clients
(specifically Cursor) doesn't support this notification. However, we
noticed that Cursor has a hidden "command" called "mcp.toolListChanged"
that allows us to trigger the tool reload manually.
5) Finally, we try to handle the remote Radon AI server in a
non-blocking fashion. The commands to the local MCP server are now never
blocked by the remote calls. Before, when initializing, we'd need to
fetch tools list from the remote server first. Now, instead of doing
that, we allow for the Radon AI server communication to happen
asynchronously. We introduce a concept of "local" and "remote" tools.
The remote tools can be loaded at any time. When an error occurs when
executing remote tool, we check if remote Radon AI server is available
and if it is not, we reset the list of remote tools (triggering tool
update notification), then retry connecting with Radon AI in the
background.
6) The above mechanism replaces the need for `ConnectionLIstener`. Now
instead of repeatedly trying to connect with the remote Radon AI server
(even when we have successfully communicated with it before), we only
check for connection issues when the remote tool is being used. We only
retry connection when the tool failed and we can no longer use MCP's
ping method to ping the server.

## Test plan
1) For most tests we use Cursor with MCP Settings page opened. It shows
the status of server and how many tools are available. Ideally we expect
Radon AI to be always shown as available, there are 4 tools in total out
of which 2 are local and 2 remote.
2) Open Radon in Cursor and notice it shows up as "extension-RadonAI" in
the MCP settings list. After initializing it briefly shows 2 tools and
then 4 tools after a few seconds.
3) We tested remote Radon AI errors by triggering Auth exceptions or
delaying the requests with timeouts. In case of auth errors we'd only
expect 2 tools to show. In case of delaying we would see 2 tools
initially and then 4 tools.
4) We tested Radon AI with wifi turned off. The server would connect and
show 2 tools although this test isn't super practical because we
wouldn't be able to use the model anyways.
5) Test turning Radon AI off in the settings. This should result in
Radon getting removed from the MCP Server page. This was tested both in
Cursor and VSCode where the logic for handling this usecase has also
been updated.
6) In VSCode after disabling and enabling RadonAI setting, the server
shows up on the list but isn't automatically started. When Copilot wants
to interact with it it starts it. But the behavior still differs
compared to the case when the extension launches on an opened project
(when project gets opened with setting ON the server starts
automatically)

---------

Co-authored-by: Ignacy Łątka <latkaignacy@gmail.com>