fix: react query in projects with non flat dependency structures (#1687)
This PR improves multiple aspects of the react-query dev tool:

1) in order to resolve issues with non flat dependency structures we
import `QueryClient` from `react-query` main package instead of
`query-core` that is only available as a dependency od `react-query` to
do do we need to move our re-export of `react-query-devtool.js` to the
end of transformed file to allow for proper initialization of
`QueryClient`

2) During testing we found that sometimes react query plugin is not
registered until the first query is performed. This should not be a
problem for most applications but we try to fix it anyway. We modify the
dev plugin registration logic so we can safely register the same module
multiple times, then we use it to try and register react query plug in
as soon as `react-query-devtool.js` logic has ended.

3) During testing we found that sometimes the files of react query might
have `.mjs` extension preventing us from detecting them in
`babel_transormer.js`, so we include them in a list of possible
`react-query` entrypoints.

fixes #1671

note: this PR has no effect on react query v3 and older as it started
using `@tanstack` module namespace since v4

### How Has This Been Tested: 

- test pnpm monorepo with expo 54 and react-qerry v5 (`5.90.5`)
- test pnpm monorepo with expo 54 and react-qerry v4 (`4.42.0`)
- test expo-52-prebuild-with-plugins with react-qerry v5 (`5.90.5`)
- test expo-52-prebuild-with-plugins with react-qerry v4 (`4.42.0`)

### How Has This Change Been Documented:

internal