feat: Add Launch configuration management and DevTools test coverage (#1686)
This PR adds test coverage for the following:  
- Launch configuration management  
- Developer Tools – Redux DevTools and React Query DevTools  

It also includes the following improvements:  
- Enhancements for screenshot and video tests – now verifies that the
saved file is not empty and not a single-color image
- Fixes for Radon AI tests (previously failing on GitHub CI)  

Also adds running the test set with DevTools tests on the `app
expo-52-prebuild-with-plugins` nightly.

### How Has This Been Tested  

Tested both locally and on GitHub Actions.  
To reproduce local tests, run one of the following commands:  
`npm run prepare-and-run-tests -- expo-52-prebuild-with-plugins 6 15 16`
`npm run run-tests-on-VM -- expo-52-prebuild-with-plugins 6 15 16`

### How Has This Change Been Documented  

No documentation changes are required.