Only verify whether launch config exists if the previous config was a custom one (#1691)
In #1684 we broke the mechanism of reloading for "detected"
configurations.

When detected configuration is the one that's currently selected, the
added method would fail to find a matching configuration and as a
consequence prevent reloading.

This PR wraps the whole logic in a check for whether the configuration
was Custom. While it makes some sense to also consider Detected
configurations for this logic, we don't track the app roots the same way
as we track configurations from the launch config, so adding this check
there wouldn't be super reliable. On top of that, the detected
configurations are now actually calculated on the UI, so we would need
to reshuffle code a bit in order to reduce coupling between the places
where we rely on that logic. Finally, deleting a "detected"
configuration isn't as easy as deleting an entry from launch.json, one
would need to remove the whole folder or the package.json from the
folder. Once someone does that they will likely get some other errors
when reloading, so this process isn't as sensible to the situation when
someone keeps using the deleted configuration.

### How Has This Been Tested: 
1. Open any projects
2. Select one of the "Detected applications" from config selector
3. Click reload -> previously you'd get an error, now it reloads
normally