Add radon-mcp package that implements stdio MCP server for Radon AI (#1670)
This PR adds a new NPM package `radon-mcp` that implements a stdio based
MCP server that proxies traffic to Radon AI http SSE server controlled
by the extension.

The main motivation is that we want to simplify the process of
integrating MCP server with different MCP clients that don'
t provide extension level programmable API. This includes tools like
Claude Code, Windsurf and more. Because we use dynamic ports for the MCP
server and the server is also tied to specific workspace, we'd need to
have the user re-add the server configuration every time they restart
the extension which isn't practical.

### The approach
1) In order to solve the above limitations, we're adding a new
`radon-mcp` command that will be available as an NPM package. It is a
common practice for MCP servers to be used via `npx` and we aim for this
command to be consumed that way. The command would start a stdio MCP
server locally that'd proxy the traffic to the actual MCP server
maintained by the extension process running in VSCode/Cursor.
2) For the command to tell the port at which it should connect with the
MCP server, we will expect the workspace folder of the project to be
provided as command argument. For example, when we want to connect with
Radon instance running for VSCode project at `~/Projects/ExpoApp`, we
pass that workspace folder as argument. Alternatively, when no argument
is provided, the command assumes that the current working directory is
the project location
3) Once the command knows the path, it locates a special file we write
into Radon's cache folders. Under that folder we keep JSON configuration
for each workspace where Radon was opened. We use a naming convention
for the config files and expect the configuration to be under
`radon-mcp-${hash}.json` where `hash` is the `MD5` of the normalized
workspace folder path (absolute path with no leading slash).
4) The stdio server reads the configuration path and attempts to connect
to the HTTP server to proxy the commands
5) In case the file doesn't exist or it can't connect, it waits for the
file being updated or for when the server is available online
6) In case of any errors when communicating with the HTTP server, the
command goes back to waiting for the server to come back online.
7) In general, we avoid shutting down the stdio server as a consequence
of errors or when disconnected from http server because MCP often never
attempt to restart the server after it's shut down on its own.

### Retries and error handling
After testing this command with some MCP client implementations, we
found out that some of them doesn't tolerate long initialization times.
Therefore the command doesn't wait with "initializing" the stdio server
for the HTTP server to connect. Instead, we perform initialization
immediately and only wait for the HTTP server when listing tools.

### How Has This Been Tested: 
1) Run npx @modelcontextprotocol/inspector providing the path to the
radon-mcp command and use list tools to list available tools
2) Start the inspector first and then start VSCode with the workspace
open -> see it go from 0 tools to 4 tools
3) Kill VSCode and observe that it again the tools doesn't appear in the
inspector.
4) Start VSCode first and then inspector -> see it loads the tools
immediately