fix: react-query plugin not detected (#1712)
On some monorepo setups, the `react-query` plugin was _still_ not
detected, because it used the `.cjs` commonjs entry point instead of one
of the files we looked for.
To avoid this issue, we relax the condition on importing our plugin to
instead look for _any_ file under the `react-query` module, and depend
on the `require` caching mechanism to ensure our hook is set up only
once.

### How Has This Been Tested: 
- ran expo-54-prebuild-with-plugins to ensure there's no regression on
our test.