fix: network inspector - fix message cleaning on LogBox trigger (#1714)
### Description

This PR fixes network messages disappearing when LogBox was triggered,
because of the `appReady` event generated. The mechanism was introduced
to clean the messages on app reload. Instead, now whenever network
plugin is initialised on application end, it sends
`IDE.clearStoredMessages` message to the extension, to achieve desired
effect.

The mechanism has to be in place in order to avoid situation where,
because we are now keeping the message information on the extension end,
the app reloads and loses information about response bodies (kept in
buffer in the app), while the messages are still available to inspect on
the panel end.

### How Has This Been Tested: 

Features were tested in local development environment of extension as
follows:
- open radon-ide/packages/vscode-extension
- [ON FIRST RUN] build simulator-server locally as follows:
- in `/packages` run: `git submodule update --init --recursive
packages/simulator-server`
  - step into simulator-server folder: `cd simulator-server`
- install dependencies as given in README.md and build
simulator-server-macos file locally using: `cargo build --release `
  - step into extension folder: `cd /packages/vscode-extension`
- run: `npm run build:sim-server-debug` or copy built binary from
simulator-server into `dist` folder
- run the extension locally using Run and Debug menu -> Run Extension
- open an application supporting Radon IDE in editor - example of app
supporting rotation: radon-ide-test-apps/react-native-80
- run Radon IDE extension panel
- *see if logbox no longer triggers the message clearing, see if the
messages are being correctly disposed on application reload*
### How Has This Change Been Documented:

Not applicable.