fix: application session is not reset when restarting app (#1725)
Fixes an issue where restarting the app doesn't reset the application
session state.
This can cause issues, for example, when the app is stopped by the
debugger and restarted (via "Restart App" or "Reinstall App" reload
actions), the app will be stuck with the debugger paused overlay despite
running underneath.