fix: step over out of user defined code (#1726)
After #1580 we no longer perform the "skip files" logic in the
`RadonCDProxyDelegate` and instead we allow the
`vscode-js-debug` to do this for us. This introduced two problems that
this PR solves:

1) the debugger controls overlay was flashing because there is no easy
way of telling that "paused" event send by the cdp is only temporary
before `vscode-js-debug` won't send back the "resumed" one. We solve
this problem by timeouting the `debuggerPausedEmitter` fire, similarly
to what we did for the "resume" event previously.
2) the `js-debug` logic was trying to repeatedly step over until it
found a file that is not part of "skip files" testing on react native
applications however showed this to generate a bad user experience, by
simply blocking the js thread for far to long. To fix this we reduce
number of possible steps from 256 to 1 in [this
PR.](https://github.com/software-mansion-labs/vscode-js-debug/pull/9).

### How Has This Been Tested: 

- run react native 82 test app 
- set a break point in `printLogs` function and step out of it 

### How Has This Change Been Documented:

internal