fix: Disable the URL bar on fatal errors (#1727)
This PR fixes a bug where the `UrlSelect` and the `Go back` button
weren't disabled when the `deviceSession` status was `fatalError`, as
the check only considered the `starting` and `running` statuses.

Correct me if I'm wrong, but I don't think this is intentional, as a
fatal error usually suggests that the simulator is completely
unresponsive, let alone to navigation.

### How Has This Been Tested: 

- Run an app
- Cause any fatal error, e.g. break some imports or disconnect the
device
- The `UrlSelect` is now disabled when the error box appears