feat: Automatically restore the connection with un- and reattached physical devices (#1733)
This PR enhances the behavior of the extension when a device got dis-
and reconnected while running an app session.

Before, when something interrupted the connection, Radon wasn't able to
restore it and a full rebuild cycle was required to continue developing.
Now, the extension "remembers" the disconnected device, and
automatically connects it to Metro, DevTools and relaunches the preview
when it becomes available again.

If the app hasn't been closed in the meantime, no refresh is needed to
restore full functionality. If anything prevented it from connecting to
DevTools, the "app disconnected from Radon" modal appears, allowing the
user to relaunch the app.


https://github.com/user-attachments/assets/5a84bae5-99cc-4cbc-83d5-5e3a6f91579f

### How Has This Been Tested: 
- Open any RN app with Radon
- Connect a physical Android device and launch the app on it
- Wait for the app to load, and then disconnect the phone/tablet
- A "device disconnected" modal should appear
- Connect the device again
- The error should disappear, and the device screen should become
visible inside Radon
- All IDE functionality should be available after a short while

### How Has This Change Been Documented:

N/A