fix: crashes when auto reload is triggered before Android device boot (#1750)
Fixes sim-server failing to start when an auto-reload is triggered
before an Android Emulator first boots.
The issue was caused by `makePreview` method being called by autoreload
(in order to restart sim-server in case of
crashes/disconnects/killswitch triggering) and accessing the
`AndroidDevice`'s `serial` property before it's initialized by the
Emulator boot procedure.
The fix checks the `serial` is actually set before accessing it, and
throwing an error in that case (instead of passing `"undefined"` to the
sim-server).

### How Has This Been Tested: 
- open Radon
- select an Android Emulator device
- trigger an autoreload immediately, before the device finishes booting
- the autoreload should succeed

### How Has This Change Been Documented:
Internal