tests: Location tests for iOS and fixes for local CI (#1759)
This pull request adds test coverage for location and localization
features on iOS devices. It also includes several fixes for tests run on
a local machine to improve overall stability.

### How Has This Been Tested:

The new tests have been executed on the local CI environment, and
previously failing tests have been verified on a locally hosted machine
to ensure they now work as expected.

To reproduce the results, the new tests should be run on the
`react-native-81-additional-libs` test application, which provides the
necessary tools for location tests.

### How Has This Change Been Documented:

Not documented yet.