fix: react query devtools circular message payloads (#1779)
React Query Devtools, on at least some versions of the library, try to
send circular objects to the devtools frontend from the application.
Because the React Native RDT implementation uses `JSON.stringify` to
serialize its messages, this causes _tons_ of errors to be thrown in the
RN app whenever a react query event that needs to be sent to the
devtools happens.
The Expo Dev Plugin for React Query works around this issue by using the
`flatted` library, which can handle (de)serialization of circular
objects.
This PR uses the same approach to resolve the problem.

### How Has This Been Tested: 
- run `social-app` in Radon
- the app should be usable and not spam `Type Error` exceptions due to
attempting to `stringify` a circular object

### How Has This Change Been Documented:
Bugfix/Internal