tests: Breakpoints test coverage (#1770)
## Summary
This pull request adds automated test coverage for the Radon IDE
breakpoints feature.

It is related to the changes introduced in this
[PR](https://github.com/software-mansion-labs/radon-ide-test-apps/pull/37)
in the test apps repository:

## How Has This Been Tested
The newly added tests were verified to work correct both on local
machine and on the CI runner.

Tests can be reproduced by running test set 4 via GitHub Actions.

## Documentation
No documentation yet.

---------

Co-authored-by: Juliusz Wajgelt <49338439+jwajgelt@users.noreply.github.com>