fix: #836 Wait for android avd ini file exists on disk. (#1805)
At my M1 Pro machine the avd.ini of the Android emulator is not always
there when the IDE needs it on booting the AVD.
Added a wait for file, which solves the stuck on booting device problem
in #836

Fixes #836 

### How Has This Been Tested: 

Ran the development version of the IDE and opened my project. This got
my Radon IDE to boot the emulator succesfully.

### How Has This Change Been Documented:

None