fix: delay before starting maestro tests (#1804)
Fixes a 30s delay after starting maestro tests and before the app is
launched on the device.
The delay was caused by `maestro` calling `applesimutils` utility in
order to set permissions on the device. The utility does not support
passing simulator device sets (and even if it did, `maestro` would not
set it anyway), so it would repeatedly try to set permissions on a
simulator device which does not exist, and timeout after 30s, after
which `maestro` would proceed with running the application to be tested.

The solution allows `applesimutils` to modify the permissions on Radon
simulator by symlinking it to the `CoreSimulator` directory -- since
`applesimutils` accesses the simulator's files directly (without
checking the simulators database, like `xcodebuild`), a symlink is
sufficient for it to function.

### How Has This Been Tested: 
- run maestro test on an iOS simulator
- the test should start soon after the Maestro TestRunner opens

### How Has This Change Been Documented:
Bugfix