feat: show explicit error when running maestro test on iPad (#1813)
Displays an explicit error when launching a maestro test on iPad
simulators.
At this time, launching maestro on iPad results in the test visibly
hanging.
This PR instead shows a vscode error dialog explaining that the feature
is not supported at the moment, and prompting the user to switch the
device in order to run the test.

### How Has This Been Tested: 
- select an iPad simulator
- try to run a maestro test
- an error should appear immediately, explaining that the test cannot be
run

### How Has This Change Been Documented:
Internal